+++ NEWS.md
@@ -2,8 +2,12 @@
 
 ## Changes in 1.1.2
 
-- `matchApply` function to allow applying any function to each match (issue
+- `lapply` and `endoapply` methods (issue
   [#84](https://github.com/rformassspectrometry/MetaboAnnotation/issues/84)).
+  `lapply` allows to apply any function to each subset of matches for each
+  `query` element and returns the corresponding `list` of results.
+  `endoapply` is similar but applies a function returning a `Matched` and
+  returns a `Matched` representing updated matches.
   
 ## Changes in 1.1.1